+++ lib/ui/chat/widgets/message_widget.dart
@@ -44,7 +44,7 @@ class MessageWidget extends StatelessWidget {
       children: [
         ChatMessageBubble(
           isSender: message.isMe,
-          color: message.isMe ? context.colors.meChatBubble : context.colors.contactChatBubble,
+          color: message.isMe ? context.colors.meChatBubble : context.colors.otherChatBubble,
           tail: !isSameSenderAsPrevious,
           child: _buildMessageContent(context),
         ),
@@ -57,7 +57,7 @@ class MessageWidget extends StatelessWidget {
               message: message,
               onReactionTap: onReactionTap,
               bubbleColor:
-                  message.isMe ? context.colors.meChatBubble : context.colors.contactChatBubble,
+                  message.isMe ? context.colors.meChatBubble : context.colors.otherChatBubble,
             ),
           ),
       ],
@@ -130,7 +130,7 @@ class MessageWidget extends StatelessWidget {
     final textStyle = TextStyle(
       fontSize: 16.sp,
       fontWeight: FontWeight.w500,
-      color: message.isMe ? context.colors.meChatBubbleText : context.colors.contactChatBubbleText,
+      color: message.isMe ? context.colors.meChatBubbleText : context.colors.otherChatBubbleText,
     );
 
     final messageContent = message.content ?? '';
@@ -321,7 +321,7 @@ class ReactionsRow extends StatelessWidget {
                               color:
                                   message.isMe
                                       ? context.colors.meChatBubbleText
-                                      : context.colors.contactChatBubbleText,
+                                      : context.colors.otherChatBubbleText,
                             ),
                             textAlign: TextAlign.center,
                           ),
@@ -334,7 +334,7 @@ class ReactionsRow extends StatelessWidget {
                             color:
                                 message.isMe
                                     ? context.colors.meChatBubbleText
-                                    : context.colors.contactChatBubbleText,
+                                    : context.colors.otherChatBubbleText,
                           ),
                           textAlign: TextAlign.center,
                         ),
@@ -351,7 +351,7 @@ class ReactionsRow extends StatelessWidget {
                                 color:
                                     message.isMe
                                         ? context.colors.meChatBubbleText
-                                        : context.colors.contactChatBubbleText,
+                                        : context.colors.otherChatBubbleText,
                               ),
                               textAlign: TextAlign.center,
                             ),
@@ -365,7 +365,7 @@ class ReactionsRow extends StatelessWidget {
                               color:
                                   message.isMe
                                       ? context.colors.meChatBubbleText
-                                      : context.colors.contactChatBubbleText,
+                                      : context.colors.otherChatBubbleText,
                             ),
                             textAlign: TextAlign.center,
                           ),
@@ -383,7 +383,7 @@ class ReactionsRow extends StatelessWidget {
               color:
                   message.isMe
                       ? context.colors.meChatBubbleText
-                      : context.colors.contactChatBubbleText,
+                      : context.colors.otherChatBubbleText,
             ),
           ),
       ],