+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 ### Changed
+- Renamed contact to user profile everywhere [#710](https://github.com/parres-hq/whitenoise_flutter/pull/710)
 ### Deprecated
 ### Removed
 - Removes metadata cache provider [#654](https://github.com/parres-hq/whitenoise_flutter/pull/654)